+++ src/test/java/org/folio/rest/support/http/ResourceClient.java
@@ -126,6 +126,10 @@ public final class ResourceClient {
       "subject types", "subjectTypes");
   }
 
+  public static ResourceClient forSubjectSources(HttpClient client) {
+    return new ResourceClient(client, InterfaceUrls::subjectSourcesUrl,
+      "subject sources", "subjectSources");
+  }
   public static ResourceClient forCallNumberTypes(HttpClient client) {
     return new ResourceClient(client, InterfaceUrls::callNumberTypesUrl,
       "call number types", "callNumberTypes");